requirements.txt
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-text-splitters
qdrant-client
tiktoken
python-dotenv
pydantic
fastapi
uvicorn
beautifulsoup4